requirements.txt
opencv-python==4.5.3.56
streamlit==0.87.0
numpy==1.21.2
Pillow==8.3.1